Convert appStateSlice to typescript
